pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "ddgs>=9.12.0",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "ipython>=9.11.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-experimental>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langchain-tools>=0.1.34",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-postgres>=3.0.5",
    "langgraph-cli[inmem]>=0.4.19",
    "python-dotenv>=1.2.2",
]